# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/url/create": {
            "post": {
                "description": "接收长链接、可选的自定义短码与过期时间，返回短码与过期时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "创建短链接",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateUrlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CreateUrlResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/{code}": {
            "get": {
                "description": "根据短码重定向到原始长链接",
                "tags": [
                    "URL"
                ],
                "summary": "短链重定向",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateUrlRequest": {
            "type": "object",
            "properties": {
                "custom_url": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "long_url": {
                    "type": "string"
                }
            }
        },
        "model.CreateUrlResponse": {
            "type": "object",
            "properties": {
                "expired_at": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                }
            }
        }
    }
}